[Release] Hero 1.4.0 with Swift 4.2 support (#534)
* Updates to swift 4.2. Fixes many linting warnings

* Tides up unnecessary code added by automatic conversion

* More tidy up.

* Update to Swift 4.2

* Update TvOSExamples for 4.2

* Update podspec for 1.4.0

* Update CHANGELOG

* Update CHANGELOG.md

* Add backwards compatability for previous veSwift rsions

* Make examples backwards compatible

* Update pods and make example compile

* Fix code review comments
