+++ .swift-version
@@ -1,1 +1,1 @@
-4.0
+4.2
