+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The changelog for `Hero`. Also see the [releases](https://github.com/HeroTransit
 
 ## Upcoming release
 
+## [1.4.0](https://github.com/HeroTransitions/Hero/releases/tag/1.4.0)
+
+### Added
+
+- Added support for Swift 4.2.
+[#534](https://github.com/HeroTransitions/Hero/pull/534) by [@rennarda](https://github.com/rennarda)
+
 ## [1.3.1](https://github.com/HeroTransitions/Hero/releases/tag/1.3.1)
 
 ### Fixed
