+++ Hero.podspec
@@ -1,6 +1,6 @@
 Pod::Spec.new do |s|
   s.name             = "Hero"
-  s.version          = "1.3.1"
+  s.version          = "1.4.0"
   s.summary          = "Elegant transition library for iOS"
 
   s.description      = <<-DESC
