+++ Sources/Debug Plugin/HeroDebugPlugin.swift
@@ -101,7 +101,7 @@ extension HeroDebugPlugin: HeroDebugViewDelegate {
     let a = CABasicAnimation(keyPath: "zPosition")
     a.fromValue = view.layer.value(forKeyPath: "zPosition")
     a.toValue = NSNumber(value: Double(to))
-    a.timingFunction = CAMediaTimingFunction(name: kCAMediaTimingFunctionEaseInEaseOut)
+    a.timingFunction = CAMediaTimingFunction(name: CAMediaTimingFunctionName.easeInEaseOut)
     a.duration = 0.4
     view.layer.add(a, forKey: "zPosition")
     view.layer.zPosition = to
@@ -165,7 +165,7 @@ extension HeroDebugPlugin: HeroDebugViewDelegate {
     let a = CABasicAnimation(keyPath: "sublayerTransform")
     a.fromValue = hero.container.layer.value(forKeyPath: "sublayerTransform")
     a.toValue = NSValue(caTransform3D: t)
-    a.timingFunction = CAMediaTimingFunction(name: kCAMediaTimingFunctionEaseInEaseOut)
+    a.timingFunction = CAMediaTimingFunction(name: CAMediaTimingFunctionName.easeInEaseOut)
     a.duration = 0.4
 
     UIView.animate(withDuration: 0.4) {
