+++ Sources/HeroViewControllerDelegate.swift
@@ -50,7 +50,7 @@ internal extension HeroTransition {
       let delegate = tabBarController.selectedViewController as? HeroViewControllerDelegate {
       closure(delegate)
     } else {
-      for vc in vc.childViewControllers where vc.isViewLoaded {
+      for vc in vc.children where vc.isViewLoaded {
         self.closureProcessForHeroDelegate(vc: vc, closure: closure)
       }
     }
