+++ Sources/Transition/HeroTransition+UINavigationControllerDelegate.swift
@@ -28,14 +28,14 @@ extension HeroTransition: UINavigationControllerDelegate {
       previousNavigationDelegate.navigationController?(navigationController, willShow: viewController, animated: animated)
     }
   }
-  
+
   public func navigationController(_ navigationController: UINavigationController, didShow viewController: UIViewController, animated: Bool) {
     if let previousNavigationDelegate = navigationController.previousNavigationDelegate {
       previousNavigationDelegate.navigationController?(navigationController, didShow: viewController, animated: animated)
     }
   }
-  
-  public func navigationController(_ navigationController: UINavigationController, animationControllerFor operation: UINavigationControllerOperation, from fromVC: UIViewController, to toVC: UIViewController) -> UIViewControllerAnimatedTransitioning? {
+
+  public func navigationController(_ navigationController: UINavigationController, animationControllerFor operation: UINavigationController.Operation, from fromVC: UIViewController, to toVC: UIViewController) -> UIViewControllerAnimatedTransitioning? {
     guard !isTransitioning else { return nil }
     self.state = .notified
     self.isPresenting = operation == .push
